Move beam profile test into BeamProfileFactoryTest
Also remove commented out code in RectangularBeamProfile.h and
add comment to clarify calculation used for beam profle in BeamProfileFactory.cpp
